requirements.txt
cycler==0.11.0
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numexpr==2.8.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
tables==3.7.0
threadpoolctl==3.1.0
torch==1.11.0
typing-extensions==4.2.0
